Employee-management/requirements.txt
amqp==5.0.5
anyascii==0.1.7
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.3.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.6
beautifulsoup4==4.8.2
billiard==3.6.3.0
cachetools==4.2.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.7.0rc1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-admin-interface==0.17.1
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-classy-tags==2.0.0
django-cms==3.8.0
django-colorfield==0.4.3
django-cors-headers==3.9.0
django-dajaxice==0.7
django-debug-toolbar==3.2.2
django-filer==2.0.2
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-formtools==2.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==4.1.1
django-modelcluster==5.1
django-mptt==0.12.0
django-polymorphic==3.0.0
django-richtextfield==1.6
django-sekizai==2.0.0
django-taggit==1.3.0
django-timezone-field==4.1.2
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
draftjs-exporter==2.1.7
drf-yasg==1.20.0
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
google-api-core==1.24.1
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.1.0
html5lib==1.1
http-ece==1.1.0
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.7.2
importlib-resources==5.1.2
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.0.1
jsonfield==3.1.0
keyring==10.6.0
kombu==5.0.2
l18n==2020.6.1
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
numpy==1.19.4
oauth==1.0.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
packaging==21.0
pbr==5.5.1
Pillow==8.0.1
prompt-toolkit==3.0.16
protobuf==3.14.0
psycopg2==2.8.6
ptyprocess==0.7.0
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==20.0.0
pyotp==2.4.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyRFC3339==1.0
pyserial==3.5
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.6.1
python-debian==0.1.32
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.4
pywebpush==1.9.4
pyxdg==0.25
PyYAML==5.3.1
redis==3.5.3
reportlab==3.4.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
reverse-geocoder==1.5.1
RPi.GPIO==0.7.0
rsa==4.6
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
schedule==1.0.0
scipy==1.5.4
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.13.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==2.0.0
telepath==0.2
testresources==2.0.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.1.2
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
virtualenv==20.4.3
wadllib==1.3.2
wagtail==2.14.1
wagtail-color-panel==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4
xhtml2pdf==0.2.5
xlrd==1.2.0
XlsxWriter==1.3.8
xlwt==1.3.0
zipp==3.4.1
zope.interface==5.2.0
